refactor cardtrack
