+++ components/CardTrack/index.ts
@@ -1,0 +1,1 @@
+export { default as CardTrack } from "./CardTrack";
