+++ components/MainTracks.tsx
@@ -1,7 +1,7 @@
 import { ReactElement } from "react";
 
 import { CardTrack, Carousel, FirstTrackContainer } from "components";
-import { CardType } from "components/CardTrack";
+import { CardType } from "components/CardTrack/CardTrack";
 import { useAuth } from "hooks";
 import { ITrack } from "types/spotify";
 import { divideArray, isCorruptedTrack } from "utils";
